Refactor dir structure of __tests-pacts__/
